Hygiene H2b: clippy clean for icelines-cli (workspace + tests)
`cargo clippy --workspace -- -D warnings` and
`cargo clippy --workspace --tests -- -D warnings` now both pass
clean. The CLAUDE.md "must be clean" bar is met.

**Targeted fixes by category** (icelines-cli has many touch-points
because it owns the CLI surface, the TUI render layer, and the test
fixtures all in one crate).

(a) **Real semantic lints fixed in production code**:
  - `commands/team.rs`, `commands/tonight.rs`: drop `&` on
    `cfg.snapshot_dir()` (needless_borrows_for_generic_args)
  - `config.rs`: collapse `if cli_no { false } else if env_off
    { false }` chains into `||`; replace `if let Some(c) = cfg.x
    { c } else { true }` with `cfg.x.unwrap_or(true)` (×2 sites,
    if_same_then_else + manual_unwrap_or)
  - `commands/analysis.rs`: hoist trailing `"Note"` literal into
    the format string (print_literal)
  - `commands/query.rs`: SortMetric → `pub(crate)` (was
    `private_interfaces` because exposed via `SortDispatch::Legacy`)
  - `tui/dashboard_panel.rs`: `len() % 2 == 0` →
    `len().is_multiple_of(2)` (manual_is_multiple_of)
  - `tui/screens/game_detail.rs`: drop 5 `as u32` no-op casts on
    fields already typed `u32` (unnecessary_cast)
  - `tui/screens/misc.rs`: `target_dates.iter().any(|d| *d == g.date)`
    → `target_dates.contains(&g.date)`; `.min(52).max(44)` →
    `.clamp(44, 52)` (manual_contains, manual_clamp)
  - `tui/screens/player.rs`: derived Default via `#[default]` on
    enum variant (derivable_impls); `sort_by(|a, b| b.cmp(&a))`
    → `sort_by_key(|s| Reverse(s))` (unnecessary_sort_by)
  - `tui/screens/team.rs`: same sort_by_key pattern
  - `tui/screens/transactions.rs`: `pub fn filter_rows<'a>(app: &'a App)
    -> Vec<&'a Transaction>` → drop the elidable lifetime
    (needless_lifetimes)
  - `tui/app.rs`: 4-clause boolean for the `'d'` shortcut extracted
    into `is_text_input_active(app)` free fn (was tripping
    nonminimal_bool + blocks_in_conditions)
  - `tui/dashboard_panel.rs`: deleted an orphan doc comment block
    that was documenting a long-removed `build_lines` function
    (empty_line_after_doc_comments)
  - `cli.rs`: replace ` + ` mid-paragraph in a clap `--chunked` doc
    comment (was parsing as a list marker → doc_lazy_continuation)

(b) **Test-only lints fixed**:
  - `config.rs`: 20× `assert_eq!(x, true)` / `assert_eq!(x, false)`
    rewritten as `assert!(x)` / `assert!(!x)` (bool_assert_comparison)
  - `config.rs`: `home_dir()` moved BEFORE the test module to fix
    items_after_test_module
  - `tui/screens/mod.rs`: 5× `app.views().iter().next()` →
    `app.views().first()` (iter_next_slice)
  - `tui/screens/mod.rs`: 2× `names.iter().any(|n| *n == "x")` →
    `names.contains(&"x")` (manual_contains)
  - `tui/screens/mod.rs`: removed a `expanded || !expanded` tautology
    from the L.5b post-fix test (overly_complex_bool_expr; my own
    leftover)
  - `tui/screens/mod.rs`: `chars().next() != Some(c)` →
    `!starts_with(c)` (chars_next_cmp)
  - `tui/dashboard_panel.rs`: `vec![…]` for a single-element
    fixture → `[…]` (useless_vec)
  - `tui/screens/goalies.rs`: 8-tuple seed type extracted into
    `type GoalieSeed<'a> = (...)` (type_complexity); dropped
    `(u32 * gp).max(0)` no-op (unnecessary_min_or_max)
  - `tui/screens/game_detail.rs`, `tui/screens/schedule.rs`:
    `#[allow(too_many_arguments)]` on synthetic test fixtures
    where struct conversion would be noisier than the flat fn

(c) **`#[allow(dead_code)]` on intentional kept items** (reserved
public surfaces, test helpers, debug-mode variants — each gets a
one-line justifying comment):
  - `Format::emit` (commands/output.rs)
  - `GroupDb::is_attended` + `AttendedGameRow::attended_at` (db.rs)
  - `PANEL_WIDTH` const + `lines_to_text` test helper
    (tui/dashboard_panel.rs)
  - `CareerTablePreset::All` enum variant (debug surface; SCOUT-6
    removed it from the keyboard cycle)
  - `format_sort_picker_row` + `run_query_views` (queries.rs;
    superseded by `_selected` / `_with_pick` variants but kept as
    public helpers)
  - `sparkline::render` (string-form helper alongside `columns`)

(d) **`#[allow(non_snake_case)]` on tests encoding literals**:
  - `parse_game_tests` mod (entire mod allowance — names literally
    encode camelCase NHL API field names: `seriesSummary`,
    `gameLabel`, etc.)
  - 4 individual fns encoding literal sentinel values: `LIVE` tag,
    `LEAGUE` placeholder team

(e) **Unused imports removed**:
  - `commands/export.rs`: drop unused `StatId` (only `StatUnit` used in scope)
  - `tui/screens/goalies.rs`: drop unused `crate::tui::headshot`
  - `tui/screens/misc.rs`, `tui/screens/mod.rs`: drop unused
    `super::*` re-exports in test modules

Workspace test count: 1297 / 0 failures — no test changes; pure
hygiene. fmt + clippy now both pass clean per CLAUDE.md.

+++ icelines-cli/src/tui/screens/game_detail.rs
@@ -209,15 +209,15 @@ fn team_leader_lines(team_label: &str, skaters: &[SkaterLine]) -> Vec<Line<'stat
         &[
             (
                 "SOG",
-                sog_leader.map(|s| (s.sog as u32, short_name(&s.player_name))),
+                sog_leader.map(|s| (s.sog, short_name(&s.player_name))),
             ),
             (
                 "Hits",
-                hits_leader.map(|s| (s.hits as u32, short_name(&s.player_name))),
+                hits_leader.map(|s| (s.hits, short_name(&s.player_name))),
             ),
             (
                 "Blocks",
-                blocks_leader.map(|s| (s.blocked_shots as u32, short_name(&s.player_name))),
+                blocks_leader.map(|s| (s.blocked_shots, short_name(&s.player_name))),
             ),
         ],
         dim,
@@ -233,11 +233,11 @@ fn team_leader_lines(team_label: &str, skaters: &[SkaterLine]) -> Vec<Line<'stat
         &[
             (
                 "Takeaways",
-                take_leader.map(|s| (s.takeaways as u32, short_name(&s.player_name))),
+                take_leader.map(|s| (s.takeaways, short_name(&s.player_name))),
             ),
             (
                 "Giveaways",
-                give_leader.map(|s| (s.giveaways as u32, short_name(&s.player_name))),
+                give_leader.map(|s| (s.giveaways, short_name(&s.player_name))),
             ),
         ],
         dim,
@@ -468,6 +468,7 @@ mod tests {
 
     /// Compact `SkaterLine` builder used by `fixture_boxscore` to keep
     /// the test fixture readable.
+    #[allow(clippy::too_many_arguments)] // synthetic test fixture; struct would be noisier.
     fn fixture_skater(
         name: &str,
         team: &str,
